+++ src/Mixin/LocalizedRouteRegistrar.php
@@ -5,12 +5,20 @@ namespace Goodcat\L10n\Mixin;
 use Closure;
 use Illuminate\Routing\RouteRegistrar;
 
+/**
+ * @mixin RouteRegistrar
+ */
 class LocalizedRouteRegistrar
 {
+    /** @var array<string, mixed> */
+    protected array $attributes;
+
+    /**
+     * @return Closure(list<string>=): (RouteRegistrar|self)
+     */
     public function lang(): Closure
     {
-        return function (array $translations = []): RouteRegistrar {
-            /** @var RouteRegistrar $this */
+        return function (array $translations = []) {
             $this->attributes['lang'] = $translations;
 
             return $this;